% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivis.R
\name{base_get_template}
\alias{base_get_template}
\title{Title}
\usage{
base_get_template(
  page_url,
  base,
  base_follow,
  target_keys = NULL,
  extract_pathes,
  req_method = "GET",
  body = NULL,
  headers = NULL,
  use_header = FALSE
)
}
\arguments{
\item{use_header}{}
}
\value{

}
\description{
Title
}
