% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivis.R
\name{chrome_addin_tutorial}
\alias{chrome_addin_tutorial}
\title{Show the manual steps to activate the required addin in Chrome.}
\usage{
chrome_addin_tutorial()
}
\description{
The identification of the correct request will be performed in a chrome extension. A dataset is generated, that will be sufficient to
build a scraper from within R.
This dataset will be transformed into a JSON object and loaded in the clipboard from where it can be accessed in R. \cr \cr
}
\details{
This function shows the steps to add the required addin to chrome.
}
