% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivis.R
\name{create_document}
\alias{create_document}
\title{Create Rmd document with scraping code}
\usage{
create_document(
  page_url,
  extract_pathes,
  response_string,
  test_eval = FALSE,
  req_method = "GET",
  Code_3_Extract = NULL,
  use_header = FALSE,
  body = NULL,
  XPathes = "",
  searchMultiPage = TRUE,
  search_multi_cols = TRUE,
  rCode = NULL
)
}
\arguments{
\item{page_url}{The url of the page ot scrape}

\item{rCode}{}
}
\description{
Sivis provides the user a script with code how to scrape the provided page.
This function uses building blocks of code to create this script.
}
\details{
The Rmd script is split in four blocks:
1. package loading and url specification
2. Performing the request
3. Extracting the content
4. Display the results
}
\seealso{
\code{\link{update_document}}
}
