% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivis.R
\name{create_run_request}
\alias{create_run_request}
\title{Create and run request to target url}
\usage{
create_run_request(cb_data)
}
\arguments{
\item{cb_data}{clipboard data. JSON from the Chrome addin shared via the clipboard.}
}
\description{
In order to find a path to extract the data from a given document, we first have to load the data in R by building a working request. We have the data that yielded a successful request from the browser. This data can not be used one to one, but is a great starting point and will yield succesful requests on
most web pages that do not prohibit web scraping.
Given this request we can download the document containg our target information.
}
