% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivis.R
\name{create_scraper}
\alias{create_scraper}
\title{Create an Rmd script with scraping code by Chrome data in the clipboard}
\usage{
create_scraper()
}
\description{
An Rmd script is created with reproducible scraping code. Input is read from clipboard
with `readClipboard()`. The clipboard is filled from the chrome addin with a json containing
target data, the page url and meta data for a working request to request this page. \cr \cr
}
\details{
The result is that an Rmd document opens. \cr
This is the top level function called by the user that initiates the process.
}
