% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivis.R
\name{evaluate_html}
\alias{evaluate_html}
\title{Get extraction path from html}
\usage{
evaluate_html(
  iter_nr,
  response_string,
  target_values,
  extract_pathes,
  xpath_from_browser,
  maxCheck,
  req_method,
  body,
  test_eval,
  use_header,
  XPathes,
  page_url,
  test_run
)
}
\arguments{
\item{iter_nr}{This function can get multiple times if the target_values have to be extracted "across multiple levels".
To avoid being caught in an infinite loop it will be stopped after n rounds. n can be set in the config.}

\item{response_string}{Character string containing the target values. Initially, it will be the response body. In follow up iterations it will be
a part of it, e.g. a json extracted from the response body.}

\item{target_values}{The strings selected by the user within the browser. For these values the corresponding extraction
method has to be found.}

\item{extract_pathes}{The target data can be nested in multiple data formats. For example a json containg html code.
To build and dynamically update the scraper all extraction steps have to be available and in the correct order.
Therefore, this extraction steps will be stored in this variable: extract_pathes and updated in the corresponding
extraction functions.}

\item{xpath_from_browser}{The xpath being generated by the browser addin. It can be helpful choosing the correct xpath if there are multiple candidates.}

\item{req_method}{The required request method to perform the request. Potentially all metods of \url{https://developer.mozilla.org/de/docs/Web/HTTP/Methods}.
Currently supported are POST and GET requests.}

\item{body}{The body of the request.}

\item{test_eval}{Parameter for automated testing}

\item{use_header}{Binary parameter (TRUE / FALSE). Should a header be used when performing a request. Note that requests can}

\item{page_url}{The url on which the data were selected. This is not necessarily the url of the request, but the url on which the content
was shown to the user in the browser.}

\item{test_run}{Parameter for automated testing}
}
\description{
Find the extraction path to extract the target values from a html. \cr
Wrapper for \code{\link{extract_html}}.
}
