% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivis.R
\name{extract_data}
\alias{extract_data}
\title{Wrapper function to create extraction pathes for html or jsons}
\usage{
extract_data(
  response_string,
  doc_type = NULL,
  cb_data,
  xpath_from_browser = "",
  body = NULL,
  extract_pathes = list(),
  page_url = page_url,
  target_values,
  iter_nr = 0,
  test_run = FALSE,
  test_eval = FALSE,
  req_method = "GET",
  headers = NULL,
  use_header = FALSE,
  req_single_quote = NULL
)
}
\arguments{
\item{response_string}{Character string containing the target values. Initially, it will be the response body. In follow up iterations it will be
a part of it, e.g. a json extracted from the response body.}

\item{doc_type}{The document type of the response_string. Initially, the document type can be read from the response headers. In follow up rounds
it will be passed as NULL and determined within this function.}

\item{cb_data}{clipboard data. JSON from the Chrome addin shared via the clipboard.}

\item{xpath_from_browser}{The xpath being generated by the browser addin. It can be helpful choosing the correct xpath if there are multiple candidates.}

\item{body}{The body of the request.}

\item{extract_pathes}{The target data can be nested in multiple data formats. For example a json containg html code.
To build and dynamically update the scraper all extraction steps have to be available and in the correct order.
Therefore, this extraction steps will be stored in this variable: extract_pathes and updated in the corresponding
extraction functions.}

\item{page_url}{The url on which the data were selected. This is not necessarily the url of the request, but the url on which the content
was shown to the user in the browser.}

\item{target_values}{The strings selected by the user within the browser. For these values the corresponding extraction
method has to be found.}

\item{iter_nr}{This function can get multiple times if the target_values have to be extracted "across multiple levels".
To avoid being caught in an infinite loop it will be stopped after n rounds. n can be set in the config.}

\item{test_run}{Parameter for automated testing}

\item{test_eval}{Parameter for automated testing}

\item{req_method}{The required request method to perform the request. Potentially all metods of \url{https://developer.mozilla.org/de/docs/Web/HTTP/Methods}.
Currently supported are POST and GET requests.}

\item{headers}{The request headers.}

\item{use_header}{Binary parameter (TRUE / FALSE). Should a header be used when performing a request. Note that requests can}

\item{req_single_quote}{Boolean parameter if the json key and values are identified with a single quote. Example:
{'price': '69'} instead of {"price": "69"}}
}
\value{
Does not return values (yet). Todo: Change it, that create document is not nested within this function.
}
\description{
#' Parent function for \code{\link{get_json_structure}} and \code{\link{extract_html}}. \cr
This function can get multiple times if the target_values have to be extracted "across multiple levels". E.g. if within
}
