% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivis.R
\name{extract_data_write_rmd}
\alias{extract_data_write_rmd}
\title{Loop through extraction until target values are found and write corresponding pathes to an rmd file}
\usage{
extract_data_write_rmd(sivis = sivis, test_run = FALSE, test_eval = FALSE)
}
\arguments{
\item{sivis}{Environment to store extraction pathes etc. Todo: should be turned into an object}

\item{test_run}{Parameter for automated testing}

\item{test_eval}{Parameter for automated testing}
}
\description{
Extract the target values from the reseponse body. Target values might be nested in multiple data formats.
E.g. target values might be in a json, that is within javascript code that is within html code. Then
the xpath will be generated to identify the location within html, next regex is created to extract the json
from the javascript code. Next, an extraction path is generated that yields the target values from the json. \cr \cr
Finally write the required extraction path in the rmd doc
}
