% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivis.R
\name{extract_html}
\alias{extract_html}
\title{Find the extraction path from html}
\usage{
extract_html(
  response_string,
  target_values,
  extract_pathes,
  xpath_from_browser = "",
  maxCheck = 5
)
}
\arguments{
\item{response_string}{Character string containing the target values. Initially, it will be the response body. In follow up iterations it will be
a part of it, e.g. a json extracted from the response body.}

\item{target_values}{The strings selected by the user within the browser. For these values the corresponding extraction
method has to be found.}

\item{extract_pathes}{The target data can be nested in multiple data formats. For example a json containg html code.
To build and dynamically update the scraper all extraction steps have to be available and in the correct order.
Therefore, this extraction steps will be stored in this variable: extract_pathes and updated in the corresponding
extraction functions.}

\item{xpath_from_browser}{The xpath being generated by the browser addin. It can be helpful choosing the correct xpath if there are multiple candidates.}
}
\description{
Find the extraction path from html
}
