% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivis.R
\name{get_json_structure}
\alias{get_json_structure}
\title{Get extraction path for a JSON object or array}
\usage{
get_json_structure(
  response_string,
  target_values,
  extract_pathes = list(),
  req_single_quote = NULL
)
}
\arguments{
\item{response_string}{character string(s) of type JSON object or array containing the target_values}

\item{target_values}{The strings selected by the user within the browser. For these values the corresponding extraction
method has to be found.}

\item{extract_pathes}{The target data can be nested in multiple data formats. For example a json containg html code.
To build and dynamically update the scraper all extraction steps have to be available and in the correct order.
Therefore, this extraction steps will be stored in this variable: extract_pathes and updated in the corresponding
extraction functions.}

\item{req_single_quote}{Boolean parameter if the json key and values are identified with a single quote. Example:
{'price': '69'} instead of {"price": "69"}}
}
\value{
A list of three variables:
\itemize{
  \item all_found: Boolean parameter. Are all target values found or should an additional extraction step by attempted.
  \item extract_pathes: See Input value.
  \item result_values: The resulting texts that are produced as output from the extraction step. In case the final
  extraction step is reached these texts equal target_values or obtain a minimal distance to it.
}
}
\description{
Get extraction path for a JSON object or array
}
\examples{
data("doc_json")
target_values <- "69"
get_json_structure(response_string = doc_json, target_values)
}
