% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivis.R
\name{get_leaf_pathes}
\alias{get_leaf_pathes}
\title{Get the pathes of potential neighbour columns}
\usage{
get_leaf_pathes(doc, tags)
}
\arguments{
\item{doc}{html document}

\item{tags}{tags of the target text selected by the user in the browser}
}
\description{
Helper function for \code{\link{get_multicol_candidates}} in which we find potential neighbour functions by:
\itemize{
  \item have the count for all target tags
  \item go up the tree (until html tag) and count amount of elements on each level.
  \item Highest level that has same amount of items than the amount of target tags is our starting point*
  \itemize{
     \item deriving the xpath of that node is our rootpath for all potential nieghbour columns
     \item all children of that starting point that dont have children themselves could contain text that qualify as potential neighbour columns**
  }
}
\strong{Exceptions:} \cr
\itemize{
  \item * a valid and useful starting point could be as well a parent node of length 1, since it will not contain additional unrelevant data.
  \item ** useful text of neighbour columns do not necessarily have to be on leaves that do not have children. Example: <div> Text <p></p></div>. \cr
}
This additional search will be performed within \code{\link{search_parent_nodes}}.
}
\examples{
data("doc_html")
doc <- xml2::read_html(html)
xpath <- "/html/body/table/tr/td[1]"
tags <- rvest::html_nodes(doc, xpath = xpath)
get_multicol_candidates(html, xpath, list())
}
