% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivis.R
\name{get_text_given_root}
\alias{get_text_given_root}
\title{Get the text of all elements that fulfill roothPath + XPath}
\usage{
get_text_given_root(root_path, XPath, doc)
}
\arguments{
\item{root_path}{a}

\item{XPath}{v}

\item{doc}{html document yielded from the server response}
}
\value{
list of texts of the elements that were identified with the XPath given the root_path + this XPath that was used.
}
\description{
Helper function to get neighbour columns of a target column. For details see: commonxpathdata
}
