% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivis.R
\name{is_html}
\alias{is_html}
\title{Test if given document is of type html.}
\usage{
is_html(response_string, target_values)
}
\arguments{
\item{response_string}{The text of the server response for which it is expected to contain the target values}

\item{target_values}{The strings selected by the user within the browser. For these values the corresponding extraction
method has to be found.}
}
\value{
Return boolean value (TRUE / FALSE) if response_string is of type html.
}
\description{
Test if given document is of type html.
}
\examples{
doc_html <- "<html><div><a>target_text</a></div></html>"
is_html(response_string = doc_html, target_values = "target_text")

}
\seealso{
\code{\link{find_doc_type}}
}
