% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivis.R
\name{mgsub2}
\alias{mgsub2}
\title{Gsub for multiple values accounting for empty strings in string and pattern}
\usage{
mgsub2(string, pattern, replacement, recycle = FALSE, ...)
}
\arguments{
\item{string}{a character vector where replacements are sought}

\item{pattern}{Character string to be matched in the given character vector}

\item{replacement}{Character string equal in length to pattern or of length 
one which are a replacement for matched pattern.}

\item{recycle}{logical. should replacement be recycled if lengths differ?}

\item{...}{arguments to pass to \code{\link[base]{regexpr}} / 
\code{\link[base]{sub}}}
}
\value{
Converted srring
}
\description{
Wrapper for \code{\link[mgsub]{mgsub}} allowing for empty strings in pattern and string parameters
}
\examples{
mgsub2("hey, ho",pattern = c("hey","ho"), replacement = c("ho","hey"))
mgsub2("", pattern = c("hey","ho"), replacement = c("ho","hey"))
mgsub2("hey, ho", pattern = "", replacement = c("ho","hey"))
}
