% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivis.R
\name{prepare_extraction}
\alias{prepare_extraction}
\title{Prepare the extraction phase with the initial data from chrome}
\usage{
prepare_extraction(sivis)
}
\description{
The function follows \code{\link{create_run_request}} and is nested within \code{\link{extract_data_write_rmd}}.
After a succesful request was performed, the data will be transformed for the transaction phase.
Target values are potentially adapted due to encoding issues, request body and headers are formatted, request
method is extracted etc.
}
\details{
#' @return
\itemize{
\item{cb_data = the data from the clipboard / addin.}
\item{req_method = Request method being used. Potentially all methods from \url{https://developer.mozilla.org/de/docs/Web/HTTP/Methods}. Currently POST and GET request are supported.}
\item{body = request body, see \url{https://en.wikipedia.org/wiki/HTTP_message_body}}
\item{xpath_from_browser = The Xpath generated by the addin in Chrome.}
\item{page_url = The request url of the target page.}
\item{target_values = The selected text by the user in the browser.}
\item{headers = The header of the request.}
\item{use_header = Binary parameter (true / false) should the request header be used or not for the request from R.}
}
}
