% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivis.R
\name{search_parent_nodes}
\alias{search_parent_nodes}
\title{Get the pathes of potential neighbour columns - Search in parent nodes}
\usage{
search_parent_nodes(leaves, root_path, doc)
}
\arguments{
\item{doc}{html document}
}
\description{
Helper function for \code{\link{get_leaf_pathes}} in which we find potential neighbour functions by:
\itemize{
  \item have the count for all target tags
  \item go up the tree (until html tag) and count amount of elements on each level.
  \item Highest level that has same amount of items then amount of target tags is our starting point
}
Then we have to make an assumption: \cr
all children of that starting point that dont have children themselves could contain text that qualify as potential neighbour columns** \cr \cr
In this function we want to relax this assumption: \cr
** useful text of neighbour columns do not necessarily have to be on leaves that do not have children. Example: <div> Text <p></p></div>. \cr
This additional search will be performed within this function.
}
