% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivis.R
\name{update_document}
\alias{update_document}
\title{Update the Rmd scraping document via user selection in Shiny.}
\usage{
update_document(
  XPathes,
  root_xpath,
  page_url,
  selectedCol,
  doc = getActiveDocumentContext()
)
}
\arguments{
\item{XPathes}{The selected xpathes from the user in the shiny app when switching from one to multiple columns}

\item{root_xpath}{The common root xpath for the XPathes parameter}

\item{page_url}{The url of the page ot scrape}

\item{selectedCol}{The selected column: Todo is this relevant?}

\item{doc}{The loaded document from page_url}
}
\description{
The user is given the option to add additional columns to the dataset via a shiny app.
After the user has chosen this columns the rmd scraping document has to be updated.
Initially the document is created via \code{\link{create_document}}. It will be called
again, just with adpated parameters.
}
\details{
The Rmd script is split in four blocks:
1. package loading and url specification
2. Performing the request
3. Extracting the content
4. Display the results

In this function block 3 is updated: Extracting the content (--> with new xpathes for additional columns)
}
\seealso{
\code{\link{create_document}}
}
